% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconfig.R
\name{CreateConfig}
\alias{CreateConfig}
\title{Create the config list for simulation from a Rosette object}
\usage{
CreateConfig(
  object,
  n.sim,
  save.sim,
  type.sim,
  mode.sim,
  n.rep,
  wt.effect,
  n.round,
  mode.rep = "bio",
  seq.shrink = 2,
  seq.depth = 200,
  pop.size = 100,
  lib.shrink = 2,
  var.shrink = 0.8
)
}
\arguments{
\item{object}{A Rosette object}

\item{n.sim}{Number of simulations}

\item{save.sim}{Directory to save the simulation data}

\item{type.sim}{Type of simulation: growth or binding}

\item{mode.sim}{Simulation model code (interger 1 to 5)}

\item{n.rep}{Number of replicates}

\item{wt.effect}{Wild-type effect (binding) or doubling rate (growth)}

\item{n.round}{Number of rounds}

\item{mode.rep}{Replicate mode: "bio" (default) or "tech"}

\item{seq.shrink}{Shrinkage factor for sequencing dispersion (times seq.disp)}

\item{seq.depth}{Sequencing depth, by default 200}

\item{pop.size}{cell population size per variant}

\item{lib.shrink}{Shrinkage factor for library dispersion}

\item{var.shrink}{Shrinkage factor for variance (var.dist)}
}
\value{
A list of config
}
\description{
Create the config list for simulation from a Rosette object
}
