% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{MapBlosumScore}
\alias{MapBlosumScore}
\title{Map AA substitution to BLOSUM90 matrix}
\usage{
MapBlosumScore(wt.vec, mut.vec, aa.code = "single")
}
\arguments{
\item{wt.vec}{vector of wildtype AA}

\item{mut.vec}{vector of mutation AA}

\item{aa.code}{"single" or "triple" amino acid coding}
}
\value{
numeric vector blosum90 score
}
\description{
Stop codon is *. If no AA substitution found code as -7.
}
