% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsummary.R
\name{PlotScoreHist}
\alias{PlotScoreHist}
\title{Plot histogram of scores in Rosette object}
\usage{
PlotScoreHist(object, test.group = FALSE)
}
\arguments{
\item{object}{An Rosette object}

\item{test.group}{If TRUE, plot histograms with test group label}
}
\value{
A ggplot object
}
\description{
Plot histograms with variant group label and/or mutant group label
}
