% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconfig.R
\name{ReadFullConfig}
\alias{ReadFullConfig}
\title{Read a full config file
(with score.df, pos.df, var.dist as data frame)}
\usage{
ReadFullConfig(file)
}
\arguments{
\item{file}{Config file path}
}
\value{
A list of config
}
\description{
Read a full config file
(with score.df, pos.df, var.dist as data frame)
}
