% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosette.R
\docType{class}
\name{Rosette-class}
\alias{Rosette-class}
\alias{Rosette}
\title{The Rosette Class}
\description{
A Rosette object is a container for inferring summary statistics from a
AssayGrowth of a Rosace experiment.
}
\section{Slots}{

\describe{
\item{\code{score.df}}{A data.frame with columns index, score, pos, wt, mut, ctrl, test, blosum}

\item{\code{pos.df}}{A data.frame with columns pos, act3, act4, act5}

\item{\code{var.label}}{label for simulation ("Neutral" + (), ("Neg"), ("Pos"), or ("Neg", "Pos")}

\item{\code{var.dist}}{A data.frame with columns label, mean, sd, count}

\item{\code{disp}}{Dispersion of raw count (sequencing step)}

\item{\code{disp.start}}{Dispersion of starting variant library}

\item{\code{rounds}}{Number of rounds}

\item{\code{project.name}}{A character string indicating the name of the project}
}}

